[IO] Do not process the streamerinfo record of a file if we read the si already
it is not necessary to deserialise the streamerinfos and interact with the
type system if we did all this work already. This is visible for example when
opening and closing several small files through a TChain in a MT environment.
With this fix we keep track of the blob of bytes representing streamer info records
that we read and in case a blob has been already processed, no work is carried out.
The hash function used is sha256, therefore collisions do not take place.