Add VDT to list of builtin targets so it builds before ROOT
The dependency of Vdt::Vdt imported target should have been
automatically added by CMake due to it being listed in the
BUILD_BYPRODUCTS of the ExternalProject_Add() command, but
not all versions of CMake work, so it is necessary to build
it early by force. Targets listed in ROOT_BUILTIN_TARGETS
get added as a dependency of the move_headers target, which
is reasonable since they often provide headers without which
ROOT cannot be built in any case (e.g. vdt/vdtMath.h).