[DF] Avoid concurrent destruction of input and output trees
At the beginning of a task a new input TChain replaced the old input
TChain for that _thread_ in TTreeProcessorMT, causing a destruction of
the previous _thread-local_ TTree.
At the same time, the initialization of a new task _in a different
thread_ could delete the output TTree of a multi-thread Snapshot
action, which was listed in the fClones of the input TTree.

This race caused the input ~TTree to access an output tree just
after its deletion, in TTree::CopyAddresses.

The root cause of this issue is that TTreeProcessorMT uses thread-local
variables while RDataFrame uses processing slots which threads can
exchange between themselves.