[DF] Use RJittedCustomColumn when jitting custom columns
In line with what it's done for jitted filters, now the jitting
of custom columns creates the RCustomColumn and assigns it as the
fConcreteCustomColumn of the corresponding RJittedCustomColumn that
was previously booked.
This avoids having the situation in which a certain custom column
has been "booked" but is not yet present in the map of custom columns.

It will also help with the "local custom columns" that are coming
with ROOT-9465.