[TMVA] CVSplit -- Improve fold assignment for large values
The fold assignment is relying on a conversion from a double to an
integer (due to using a TTreeFormula). It was previously enforced that
the formula output should be "close" to an integer. The semantics are
now changed to assume the output is an integer and round any potential
deviation to the closest one.
This should ensure that the valid range is increased significantly.