[cxxmodules] Fall back to Header parsing on demand if we have no module.
We assumed that we will always have module file and unconditionally disable
header parsing on demand. However, the major use-case is gradual migration
to modules. In this scenario (tested by root-meta-fwdDecls-fwdDeclarations),
we have a dictionary which has no module file and still relies on the old
behavior. This can realistically happen when users gradually migrate to
modules. For example, we have modules-aware ROOT and untouched third party
code.

This patch enables header parsing on demand when we have no module file
available.